1. pif_step_motor 관련 ㅏ일에 오류를 수정함.
   - resolution 타입 불일치
  - stage_index 범위 검사 누락

+++ include/motor/pif_step_motor_pos.h
@@ -69,7 +69,7 @@ extern "C" {
  * @param period1ms Control task period in milliseconds.
  * @return `TRUE` on success, otherwise `FALSE`.
  */
-BOOL pifStepMotorPos_Init(PifStepMotorPos* p_owner, PifId id, PifTimerManager* p_timer_manager, uint8_t resolution,
+BOOL pifStepMotorPos_Init(PifStepMotorPos* p_owner, PifId id, PifTimerManager* p_timer_manager, uint16_t resolution,
 		PifStepMotorOperation operation, uint16_t period1ms);
 
 /**
