+++ include/motor/pif_step_motor_speed.h
@@ -67,7 +67,7 @@ extern "C" {
  * @param period1ms Control task period in milliseconds.
  * @return `TRUE` on success, otherwise `FALSE`.
  */
-BOOL pifStepMotorSpeed_Init(PifStepMotorSpeed* p_owner, PifId id, PifTimerManager* p_timer_manager, uint8_t resolution,
+BOOL pifStepMotorSpeed_Init(PifStepMotorSpeed* p_owner, PifId id, PifTimerManager* p_timer_manager, uint16_t resolution,
 		PifStepMotorOperation operation, uint16_t period1ms);
 
 /**
